Only create school moves for patients currently in school during class import
Class imports have a post-processing step to create school moves for every patient in the school, but not in the uplaoded file.
The previous logic used PatientLocation to determine whether a patient is currently in the school or not. This is problematic when the patient is in the session for the school but not in the school, which can happen when the patient has been moved elsewhere after being vaccinated for example. This would falsely create a school move from the current school to "unknown school" when a class import is done for the school the patient was previously in.

MAV-2648