Handle re-reviews when only change is school moves out of school
In the case none of the records in the file need re-review, but data has changed so that more school moves out of the school need to be created, we must put the import back into review. This prevents an inconsistent state where patients that are not in the file are still in the school, wither when partially or completely processed.

MAV-2464