Touch user sessions when calling from reports app
When a user logs into Mavis, they create a Warden session. Every time
they make a request to Mavis, the Warden session is refreshed in the
database.

When they go to the reporting app, we create a JWT on their behalf,
which is used on subsequent API requests.

However, when an API request is received, it does not refresh the Warden
session.

So if a user spends too much time (>15.minutes) in the Reporting app,
even if they're interacting with it, eventually when they return to
Mavis they will be signed out.

Now, there is no built-in `Devise.touch_session(session_id, user)`. And
Warden doesn't offer an easy way to refresh a user's session unless
you're in a request lifecycle. Our API requests are being sent from the
Python backend, so they can't access the user's Mavis cookies.

However, we can query `ActiveRecord::SessionStore::Session` and edit the
timestamp that Warden relies on directly in the database.

This feels a bit like we're breaking their abstraction, but actually
what Devise is doing is very similar in `devise/hooks/timeoutable.rb`:

```rb
warden.session(scope)['last_request_at'] = Time.now.utc.to_i
```