Add vaccine criteria parameters
This adds two methods on the `VaccineCriteria` method to be able to convert
to and from a string value which we can use to determine which status to
show. The programme is included in the string as we need to be able to
filter and show custom sub-statuses per programme.

Jira-Issue: MAV-2657