Ensure seeds run without Redis
In fad2ee4001b073343534e5371b548f15f05d19e1 a callback was added to the
`PatientTeam` model which tries to queue a job. When seeding the
database Redis may not necessarily be available and therefore we need a
way of skipping this callback action.

Unfortunately callbacks are difficult to configure/skip so I've had to
introduce an attribute on the model that is only used when seeding the
database.

This should only be temporary until we are able to refactor
`PatientTeam` to avoid using callbacks.