Improve search by name for locations
# Shortcomings of the old approach

The previous hybrid approach between ILIKE and SIMILARITY was problematic. 
Certain names would match for as long as the name was less than 3 characters
but would no longer match after the name surpassed that length because of the 
long name of the school not having high enough similarity to the search query.

# Filtering in the new approach
Whereas similarity requires that the entire name matches the entire query,
word similarity only requires that the entire query matches some part of the 
name. So "Co" will be an excellent match to any name that contains a word 
starting with "Co". This means that word similarity behaves similar to our old 
search except in a handful of cases where it includes additional results that
would otherwise be missed due to the school name being so long that the 
similarity did not yet exceed 0.3 after a reasonable number of characters was 
typed.

# Ordering of the results
For ordering, we still use a regular similarity, rather than word similarity. 
This is to ensure that when searching for "St. John's School", you will find
"St. John's School" before finding "St. John's School for Girls".

# Performance of this new search
I have used EXPLAIN analyze to analyze the new query and it performs similar to 
the old query. Since no indices are present, a sequential table scan is
performed regardless of the query used. We have not received any reports that 
the location search is currently too slow and also have no reason to expect
significant changes in performance since all schools are already stored in the
database and no new schools will be added for new teams (in this locations 
differ significantly from patients, where new patients get added with every new
team onboarded).

If ever performance improvements were required, we could add a trigram index to 
the name column and use the boolean operator %> instead of the word similarity 
function. I chose not to do that for now as it would require 
pg_trgm.word_similarity to be set to a different value which adds needless 
complexity.

Jira-Issue: MAV-2680