Improve search by name for locations
# Shortcomings of the old approach

The previous hybrid approach between ILIKE and SIMILARITY was problematic. 
Certain names would match for as long as the name was less than 3 characters
but would no longer match after the name surpassed that length because of the 
long name of the school not having high enough similarity to the search query.

# Filtering in the new approach
The new approach does away with the hybrid setup and only uses WORD_SIMILARITY,
through the %> operator. 

Whereas similarity requires that the entire name matches the entire query,
word similarity only requires that the entire query matches some part of the 
name. So "Co" will be an excellent match to any name that contains a word 
starting with "Co".

One notable regression from the old search is that single characters no longer 
match to any results except when that single letter would be the full name. 
This may affect autocomplete suggestions.

# Ordering of the results
For ordering, we still use a regular similarity, rather than word similarity. 
This is to ensure that when searching for "St. John's School", you will find
"St. John's Primary School" before finding "St. John's School for Girls with 
learning difficulties". Essentially it favours results that are matched 
entirely by the query over results that merely contain the query.

# Performance of this new search
I have used EXPLAIN analyze to analyze the new query and it is about 3x slower 
than the old query was. Since no indices are present, a sequential table scan is
performed regardless of the query used, but performing a TRANSLATE adds 
significant computational time, and WORD_SIMILARITY is slightly less efficient 
than SIMILARITY itself. That said, on my local machine it is still possible to 
search through all 52000 locations that exist in GIAS in 300ms, and in 
production, teams will not search through all 52000 schools but only those that
belong to their own team.

If ever performance improvements were required, we could add a trigram index to 
the TRANSLATE(locations.name, '.''', '') and 
TRANSLATE(locations.alternative_name, '.''', '').

Jira-Issue: MAV-2680