Accept more `-immunization.target` formats
The API team recently introduced another bug in their `Bundle.link`,
(presumably in the process of fixing their old bug), where they
now use `immunization-target`. This has only been seen in their
staging environment, and we have notified them, so hopefully this
never reaches production. In the meantime we need to be able to
test and demo against their INT environment, and need to be robust
against this bug in case it makes its way into prod.

Jira-Issue: MAV-