Don't count nasal preferred consent as injection
This fixes an issue where in the flu session overview boxes, patients
with consent for both nasal and injection would appear under the "Due
injection" statistics box which is incorrect as they are due nasal
instead. This meant that clicking on the "Due injection" statistic would
result in seeing no results.