Add single-page vaccination report form
VaccinationReportsController currently implements a standard, one thing
per page form that guides users through choosing the programme, dates,
and file format for their desired report.

We have evidence from our users that they would prefer a quicker
interaction, as the current approach requires significant back and forth
when downloading data across programmes.

This adds a `/vaccination-report/new` route which is not currently
linked to from anywhere, to aid with initial testing. We considered
adding a feature flag, but given the simple nature of the form and that
it's obfuscated, it should be low-risk to merge without one.

This approach also allows deep-linking from the reporting app to the
first step of the form, which isn't possible currently, since the
existing form starts with a `POST` to specify the programme and academic
year context.

Once it passes testing, we can swap out the existing Wicked
implementation, removing the templates and actions.