Don't retry on `NHS::PDS::InvalidSearchData`
This ensures that we don't end up with jobs stuck in the retry set which
are never going to fail, and are likely to have already been manually
matched.

In the future, this job is going to be responsible for sending out
confirmation emails as we need to know the patient's year group to know
which session they will go to. This commit makes the job safer by adding
an expected scenario and handling it explicitly.

Jira-Issue: MAV-2663