Add `Notifier::Consent`
This adds a new class which is responsible for sending notifications
related to consent and triage, replacing the `TriageMailerConcern`. The
reason behind replacing the concern is that we want to be able to send
notifications from various different places, and the concern is
described as a controller concern, which is not where it is used.

Jira-Issue: MAV-2663