Rename `ConsentFormMatchingJob`
This job will soon be responsible for more than just matching the
consent form with the patient, so I'm renaming it to make it more
generic.

To rename this job safely, the old class is kept around in case there
are any jobs on the queue when the deploy happens.

Jira-Issue: MAV-2663