Track when consent form confirmations are sent
This adds a new column to the consent form table,
`confirmation_sent_at`, which will be used to track when the consent
form confirmation is sent.

We need this as a safety mechanism to avoid parents receiving the
confirmation email and text multiple times when the responsibility of
sending this notification is moved to a background job, which has the
potential to be retried.

For now this change is just to make it such that the value is tracked.

Jira-Issue: MAV-2663