Ensure invite to clinic results in a programme status
This fixes a bug where the triage status of `invite_to_clinic` was
resulting in an overall programme status of `not_eligible` which was
incorrect. Instead the status should be `needs_triage` to reflect the
fact that triage will need to happen again before the patient can be
vaccinated.

Arguably it might make sense to introduce a new programme status for
this, but that is something we can add later.

This was found as some of the tallies weren't adding up as these
patients were being excluded from any counts.

Jira-Issue: MAV-2689