Fix bug related to patient teams being deleted
This fixes an issue where patient teams were being deleted incorrectly
if one of many contributing rows was deleted or updated.

The bug manifests itself when there exists only 1 type of reason why a team
should be able to access the patient, and within that category there are
multiple instances, and either one of those instances gets updated or deleted.

For example, let's say a patient has been added to two sessions for a team,
and that's their only relationship to the team (say school session and clinic
session). If the patient is then removed from the school session, their
relationship to the team should stay the same because they still exist in the
clinic session. However, the bug was causing the clinic session association to
be ignored, and the patient would be considered no longer part of the team.

Jira-Issue: MAV-2713