Track original session on consent forms
This adds a new reference on the consent form model for tracking the
original session that the consent form was created from.

We need this because there now might be multiple sessions for the same
programme and location, but for different year groups, and since we
don't ask for year group on the form we need to know which session the
consent form was originally sent from, to be able to know which dates to
give the parents.

Jira-Issue: MAV-2663