Hide cancel button on import review screen in some circumstances
If the only change to trigger a re-review flow for a class import is that there are new school moves from the relevant school to "Unknown school", then the user doesn't really have a choice in accepting or rejecting this as it is part of the post processing step for class imports. We only trigger the re-review (as of MAV-2646) so that the user is aware new school moves will be created. Since they cannot reject the creation of these school moves, we should hide the "Ignore changes" button.

MAV-2656