Use original session if possible
This updates how the session is determined for a consent form, which is
then used on the confirmation page to show which dates the patient is
likely to be invited to.

If the parent has said their child doesn't go a different school, we can
assume the consent form was sent correctly and we can show the right
dates.

Jira-Issue: MAV-2663