Add `vaccination_and_dates` personalisation
This adds a new personalisation variable which includes both the
vaccination that will be given and the dates. We need this to support
future content on emails and texts where the dates may not be shown
because they're deemed to not be accurate enough.

By making this change now, we can modify the templates to start using
this personalisation variable, and then modify the value later without
needing to change the templates.

Jira-Issue: MAV-2663