Hide session dates if not accurate enough
Previously we assumed all year groups would be administered in any
particular session, but now we support the ability for sessions to
administer programmes to different year groups, and therefore we need
the ability to know the year group of the patient before we tell the
parents which dates their child will be vaccinated in.

To support this, we will first attempt to match the consent form with
the child, and we introduce some logic to determine whether the session
we've identifier for this consent form is accurate enough to display
the dates for.

Jira-Issue: MAV-2663