Expire triage delay vaccination on day instead of day after
This fixes an issue where the delay triage would end up expiring the day
after the date that the nurses have specified as the earliest date a
patient can be vaccinated again.

For MMR, this meant that patients would have to wait 29 days instead of
the correct 28 days.

Jira-Issue: MAV-2696