Don't require sessions for patient teams
This updates the logic around when patient-team associations are created
to avoid needing sessions to exist for a patient location to be
considered the reason for a patient-team association to exist.

We need this because in the future teams will start with no sessions,
but patients will exist in their locations and therefore we need the
ability for teams to view those patients.

Jira-Issue: MAV-2645