Don't add PDS response to breadcrumbs
This can result in PII being leaked via the issue, instead we should
investigate issues like this by repeating the request manually and
seeing the response.

We could try to filter out the response details, but it's difficult to
get that exactly right and for now it'll be easier to avoid sending it
in the first place.

Jira-Issue: MAV-2727