Remove unused SIDEKIQ_CONCURRENCY
Technically setting the database pool to Rails threads + Sidekiq threads isn't
correct. It was when GoodJob was running inside the Rails process
asynchronously, but that is no longer the case and we only want as many
connections in the pool as we have workers, be they Rails or Sidekiq.

Jira-Issue: MAV-2744