Increase location matching threshold
In a recent change, location search was reworked to allow searching for 
partial names. In order to remain consistent with previous behaviour, the 
search was given an exceptionally low similarity threshold (of 0.3) so that it 
would remain equally forgiving of typos as before. In testing this behaviour
was found to be not desirable. Using the built in operator %> means we will now
use the pg_trgm default word similarity threshold. 

A new test was added to make sure we are not too lax with matching. 

A bug was found in the tests caused by lazy loading of let declated variables.
This was resolved and some tests were amended to accept the new results, as 
they were considered to be reasonable results. 

Jira-Issue: MAV-1093