Remove Flipper methods for MMRV vaccine feature flag
The disable/enable Flipper methods were added to ensure the rake
task `vaccines:seed` is able to load the health questions for the
MMRV vaccine.

Since we've temporarily removed MMRV vaccines from the config until
the `disease_types` column is added and populated, these methods are
no longer needed.

They can be re-added when we're ready to load MMRV vaccines in a
future deployment.