Ensure MMRV health questions get created due vaccine:seed
Since MMRV will not be going live until Jan 2026, we have to enable
and disable the feature flag when running the rake task vaccine:seed.

This is because `Programme.variant_for` is guarded by Flipper.enabled?(:mmrv).
If it's not enabled during the rake task, then `vaccine.programme.name`
will return "MMR" instead of "MMRV".