Add `ClinicNotification`
This is going to be used to represent clinic invitations (and potential
clinic booking in the future). It will replace the `SessionNotification`
model to allow for clinic invitations to not be associated with a
specific session, ensuring we can avoid duplicate invitations being
sent.

Effectively, this model will work closer to the `ConsentNotification`
model which allows for consent requests to be sent and tracked to avoid
duplicates.

Jira-Issue: MAV-2799