Add alternative name to locations
This adds a new column, `alternative_name`, to the locations table which
we can use to capture any variations of a location name that enables
locations to be found with different names.

Specifically, this will be used initially for the generic clinic locations
which have different names depending on where they are shown.

Jira-Issue: MAV-2803