Enable some feature flags in development
These feature flags are on in production and they represent the most up
to date version of the service. They are also the version that the end
to end tests run against.

This updates the seeds to enable these flags by default simplifying
development and it can be used by the end to end tests to set up a
suitable environment.

Co-Authored-By: Brage Gording <brage.gording@tngtech.com>
Co-Authored-By: Joshua Frost <joshuafrost.work@gmail.com>