Refactor `AppPatientSessionSearchResultCardComponent`
This refactors the component to be more dynamic in terms of how various
elements are shown, to make it easier to change this according to
feature flags.

Specifically, this fixes a few issues with the previous approach where
the `context` was set to `register` on the "Children" tab when the
schools and sessions feature flag was enabled. While this worked for
the most part, it lead to odd behaviours occasionally.

Jira-Issue: MAV-2645