Add `GenericClinicSessionFinder`
This replaces the `generic_clinic_session` method on the teams, because
we will shortly be adding the ability for teams to create multiple
clinic sessions.

For now this will work the same as the existing set up, because it will
always return the same session, but it prepares us for when teams have
the ability to create multiple clinic sessions.

Jira-Issue: MAV-2273