Refactor vaccination records generator cli
These were broken as a result of recent data refactoring, and also contained
bugs from the initial implementation.

For example when calculating if there were enough patients to record
vaccinations, it only considered one session. So if the user requested 5000
vaccinations, and the first qualifying session didn't have 5000 patients that
could be vaccinated in it, it would raise an error. This has been fixed to
consider all the sessions that qualify.