Add decision to upload issues
This change adds the UI for users to select a resolution option for close matches from the import review screen. We also validate that all close matches have a decision before committing the changesets from the import.

Decisions are saved on the changeset, in review_data to be used in the CommitPatientChangesetsJob to apply the appropriate action based on decision (future commit).