Marginal gains to human_enum_name performance
We noticed in recent profiling runs of vaccination exports that these would
provide some easy, if marginal, gains. The `pluralize` change seems to save
about 0.5% of the time spent to generate our test report, but the
`human_enum_name` saved closer to 2-2.5%.

The `human_enum_name` cacheing seems safe since the I18n localisations don't
change much, and the number of enum names that will be looked up is limited to
the number of enum values in the system.