Add `disease_types` column to consents table
This new column will help implement a new feature where a
nurse can get consent for an MMR vaccine even though the
patient is meant to get the new MMRV variant.

We already use the of concept of disease types in the Vaccine
model so we can distinguish between MMR and MMRV. Introducing
it into the Consent model will have the same purpose.

Jira-Issue: MAV-2811