Add logic to handle user decision for close matches
This adds the logic requrie to resolve close matches within the CommitPatientChangesetsJob itself, so no upload issues will be staged for review post-import.
We use the same logic in PendingChangesConcern treating the changeset as the object with changes. Does the change in memory so bulk import can still be performed, for better performance.