Ensure `ignored_record_count` is set for existing records
This fixes an issue where after adding the `ignored_record_count` column
to `ImmunisationImport`, all the existing records are invalid due to the
value being `nil`.

Here I'm updating the migration that has already run in QA, test and
performance, but not yet in production. I've manually run this in those
environment for now.

Jira-Issue: MAV-2850