Add `disease_type` enum type
This adds a new enum type that will be used to describe disease types
that we'll use in various places. It's designed to work in a similar way
to `programme_type`.

This PR will be followed up with PRs which adds `disease_types` columns
to various tables, and converts the existing `disease_types` column
on the `vaccines` table to use this enum.