Add indexes on `consent_forms`
This adds two indexes on the `consent_forms` table to provide fast
lookups for two common scopes, `recorded` and `unmatched`.

The `recorded` scope is used to search for consent forms which have
been submitted by a parent, and the `unmatched` scope is used to find
consent forms that haven't been matched with a patient yet.

Jira-Issue: MAV-2866