Nurse can get consent for MMR for an MMRV eligible patient
The consent journey for MMRV has been updated to handle the
scenario where MMRV vaccines have ran out of stock during a
session and the nurse still wants to offer a fallback to MMR.

The MMRV consent journey will ask if the nurse if the patient
can be vaccinated with an MMRV vaccine. If the reply is no, then
the consent journey wll proceed as using the MMR vaccine.

Jira-Issue: MAV-2811