Add task to delete sessions without dates
This allows us to migrate the existing teams over to using the new model
where we won't create sessions automatically, and instead teams are
expected to create their own sessions on an individual basis.

This is a script that we only need to run once so I've written it as a
Rake task, and already run in QA and test.

Jira-Issue: MAV-2741