Fix back link in session creation
This fixes an issue where pressing back would get the user stuck in a
loop. This was attempted to be fixed in
9fe424425a1446b6a7d66e1fe7849b1d8ae20e4b but then broke again in
f3f688c12d455238263f8479bb3b5ee3416b5aa7.

I've also removed the "Back" link from the bottom of the dates page as
this doesn't exist in the prototype and is already present at the top of
the page.

Jira-Issue: MAV-2741