Allow importing MMRV vaccination records without a vaccine
This adds support for uploading vaccination records for MMRV where the
vaccine is unknown. To support this, a new `MMRV` programme name has
been added and that corresponds to the `MMRV` programme variant.

The code in this commit is quite hard coded to MMRV and won't work for
any hypothetical future programme variants for other programmes. We'll
likely need to refactor this in the future if that becomes a need.

This is also missing a feature test which can be added later.

Jira-Issue: MAV-2814