Clarify logic of `open_for_consent?`
This renames the method `open_for_consent?` to `can_receive_consent?` to
better represent the logic in this method, and then introduces a new
`open_for_consent?` method that returns `false` if the consent window is
not open yet.

For now this will only be used to decide whether the year groups of the
session can be edited.

Jira-Issue: MAV-2741