Refactor `where_programme` scope
The `where_programme` scope was modified to accept a list of disease
types to be able to filter on particular programme variants. However,
the disease types can now be sourced directly from the `programme`
instance that is passed in to the scope. For example, if an MMRV
programme variant is passed in to the method, the disease types will
include varicella, and therefore be filtered approriately.

I've also renamed the scope to `for_programme` to match the naming
convention we have on other scopes like this.