Default `disease_types` from programme on creation
When creating `Consent` records from a `ConsentForm`, populate the
`disease_types` attribute using the default values from the
associated programme.

This is a preparatory step for adding a `NOT NULL` constraint to
the `disease_types` column on the `consents` table. By ensuring
this field is populated during the consent creation process, we
can safely apply the database constraint in a future migration
without causing data integrity errors.

Note:
`disease_types` will be added to `ConsentForm` when parents have
give consent for MMRV, for now we can default it to the programme
disease types