Default `disease_types` from programme in `DraftConsent`
When a `DraftConsent` is written to a `Consent` object, populate the
`disease_types` attribute from the associated programme's defaults if
it is not already present.

This guarantees that consents have the `disease_types` field set. It is
a preparatory step for adding a `NOT NULL` constraint to the column in
the database.