Use prefix from resolved status
This fixes an issue where the prefix shown for a resolved status (the
programme name) doesn't return the correct variant in some cases.

This works by getting the `PatientStatusResolver` class to determine the
prefix from the programme of the status model, which will be found using
the `disease_types`.

Jira-Issue: MAV-2898