Fix issue where wrong vaccine variants were chosen
This fixes an issue where the wrong vaccines for MMR and MMRV were shown
to users, because we were fetching the programmes for a patient in a
session based on their eligibility by date of birth and not via their
programme status disease types.

This change also means that the programmes when shown in the context of
a list for a particular patient will show as "MMR(V)". This works better
in terms of the design because some patients will have a mixture of MMR
and MMRV historical vaccination doses.

Jira-Issue: MAV-2902